pyproject.toml
[tool.poetry]
name = "pyamp"
version = "0.1.0"
description = ""
authors = ["acbchiella <acb.chiella@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^9.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
